+++ CHANGELOG.md
@@ -73,11 +73,23 @@ Added:
   - `Processor.verify`: handle fileGrp cardinality verification, with default implementation
   - `Processor.setup`: to set up processor before processing, optional
 
-## Unreleased
+## [2.68.0] - 2024-08-23
 
 Changed:
 
   * ocrd_network: Use `ocrd-all-tool.json` bundled by core instead of download from website, #1257, #1260
+  * :fire: `ocrd network client processing processor` renamed `ocrd network client processing run`, #1269
+  * `ocrd network client processing run` supports blocking behavior with `--block` by polling job status, #1265, #1269
+
+Added:
+
+  * `ocrd network client workflow run` Run, optionally blocking, a workflow on the processing server, #1265, #1269
+  * `ocrd network client workflow check-status` to get the status of a workflow job, #1269
+  * `ocrd network client processing check-status` to get the status of a processing (processor) job, #1269
+  * `ocrd network client discovery processors` to list the processors deployed in the processing server, #1269
+  * `ocrd network client discovery processor` to get the `ocrd-tool.json` of a deployed processor, #1269
+  * `ocrd network client processing check-log` to retrieve the log data for a processing job, #1269
+  * Environment variables `OCRD_NETWORK_CLIENT_POLLING_SLEEP` and `OCRD_NETWORK_CLIENT_POLLING_TIMEOUT` to control polling interval and timeout for `ocrd network client {processing processor,workflow run`, #1269
   * `ocrd workspace clone`/`Resolver.workspace_from_url`: with `clobber_mets=False`, raise a FileExistsError for existing mets.xml on disk, #563, #1268
   * `ocrd workspace find --download`: print the the correct, up-to-date field, not `None`, #1202, #1266
 
@@ -2220,8 +2232,12 @@ Fixed
 Initial Release
 
 <!-- link-labels -->
+<<<<<<< HEAD
 [3.0.0a2]: ../../compare/v3.0.0a2..v3.0.0a1
 [3.0.0a1]: ../../compare/v3.0.0a1..v2.67.2
+=======
+[2.68.0]: ../../compare/v2.68.0..v2.67.2
+>>>>>>> @{-1}
 [2.67.2]: ../../compare/v2.67.2..v2.67.1
 [2.67.1]: ../../compare/v2.67.1..v2.67.0
 [2.67.0]: ../../compare/v2.67.0..v2.66.1